e_commerce/requirements.txt
fastlite
faststripe==2025.05.28.1
python-fasthtml
stripe